created type file for the http api
